refactor: cast `totalDamage` to float in character classes to ensure type consistency

+++ core/src/main/java/io/github/com/ranie_borges/thejungle/model/entity/characters/Hunter.java
@@ -39,7 +39,7 @@ public class Hunter extends Character {
         logger.info("{} strikes {} with {} damage.", getName(), creature.getName(), totalDamage);
 
         float creatureLife = creature.getLifeRatio();
-        creatureLife -= totalDamage;
+        creatureLife -= (float) totalDamage;
         creature.setLifeRatio(Math.max(0, creatureLife));
 
         return creatureLife <= 0;
