+++ core/src/main/java/io/github/com/ranie_borges/thejungle/model/entity/characters/Survivor.java
@@ -39,7 +39,7 @@ public class Survivor extends Character {
         logger.info("{} attacks {} causing {} damage.", getName(), creature.getName(), totalDamage);
 
         float creatureLife = creature.getLifeRatio();
-        creatureLife -= totalDamage;
+        creatureLife -= (float) totalDamage;
         creature.setLifeRatio(Math.max(0, creatureLife));
 
         return creatureLife <= 0;
